Fix CFG generation for switch expressions
The generated CFG for switch expressions previously included the
switch expression twice, once as a terminator (as it should be) and
once as an expression in the block that used the result of the switch
expression. This would make the switch expression the top of the stack
instead of the result and once the user popped the switch expression,
the result of the switch expression would be left on the stack when
the stack should either be empty or contain the results of other
subexpressions.